requirements.txt
alabaster==0.7.9
appnope==0.1.0
arrow==0.8.0
awesome-slugify==1.6.5
Babel==2.3.4
binaryornot==0.4.0
chardet==2.3.0
click==6.6
configparser==3.5.0
cookiecutter==1.4.0
coverage==4.2
decorator==4.0.10
defusedxml==0.4.1
Django==1.9.9
django-allauth==0.27.0
django-autoslug==1.9.3
django-bootstrap3==7.0.1
django-braces==1.9.0
django-cors-headers==1.1.0
django-coverage-plugin==1.3.1
django-crispy-forms==1.6.0
django-debug-toolbar==1.5
django-environ==0.4.0
django-extensions==1.7.4
django-model-utils==2.5.2
django-multiselectfield==0.1.4
django-oauth-toolkit==0.10.0
django-redis==4.4.4
django-test-plus==1.0.15
djangorestframework==3.4.3
djangorestframework-jwt==1.8.0
docutils==0.12
factory-boy==2.7.0
fake-factory==0.7.2
future==0.15.2
imageio==1.6
imagesize==0.7.1
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
Jinja2==2.8
jinja2-time==0.2.0
MarkupSafe==0.23
moviepy==0.2.2.11
mysqlclient==1.3.7
numpy==1.11.2
oauthlib==1.0.3
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.3.1
poyo==0.4.0
prompt-toolkit==1.0.7
psycopg2==2.6.2
ptyprocess==0.5.1
py==1.4.31
Pygments==2.1.3
PyJWT==1.4.2
pytest==3.0.2
pytest-django==2.9.1
pytest-sugar==0.7.1
python-dateutil==2.5.3
python3-openid==3.0.10
pytz==2016.6.1
redis==2.10.5
regex==2016.9.22
requests==2.11.1
requests-oauthlib==0.7.0
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
sorl-thumbnail==12.3
Sphinx==1.4.5
sqlparse==0.2.1
termcolor==1.1.0
tqdm==4.10.0
traitlets==4.3.0
Unidecode==0.4.19
wcwidth==0.1.7
Werkzeug==0.11.11
wheel==0.29.0
whichcraft==0.4.0
whitenoise==3.2.1
